+++ nix/tests/default.nix
@@ -85,6 +85,13 @@ let
       drv = rustPassthru cloud-hypervisor;
       options = buildtimeOptions;
     }
+
+    # Multiple src urls
+    {
+      name = "kexec-tools";
+      drv = kexec-tools;
+      options = { };
+    }
   ];
 
   cycloneDxVersion = "1.5";
@@ -99,11 +106,12 @@ let
   buildBomAndValidate =
     drv: options:
     pkgs.runCommand "${drv.name}-bom-validation" { nativeBuildInputs = [ pkgs.check-jsonschema ]; } ''
+      sbom="${buildBom drv options}"
       check-jsonschema \
         --schemafile "${cycloneDxSpec}/schema/bom-${cycloneDxVersion}.schema.json" \
         --base-uri "${cycloneDxSpec}/schema/bom-${cycloneDxVersion}.schema.json" \
-        "${buildBom drv options}"
-      touch $out
+        "$sbom"
+      ln -s $sbom $out
     '';
 
   genAttrsFromDrvs =